refactoed q10 to be double, among other minor refactorings

+++ C#/FruitREST/Model/FoodDTO.cs
@@ -1,7 +1,7 @@
 ﻿namespace FruitREST.Model
 {
 
-    public record FoodDTO(string name, int foodTypeId, string apiLink, byte spoilDate, byte spoilHours, double idealTemperature, double idealHumidity,  string foodTypeName = "", byte q10Factor = 3, double maxTemp = 40, double minTemp = -15);
+    public record FoodDTO(string name, int foodTypeId, string apiLink, byte spoilDate, byte spoilHours, double idealTemperature, double idealHumidity,  string foodTypeName = "", double q10Factor = 3, double maxTemp = 40, double minTemp = -15);
